+++ lib/src/main/java/com/smileidentity/compose/SmileIDExt.kt
@@ -51,7 +51,7 @@ import java.net.URL
  * front camera will be used.
  * @param showAttribution Whether to show the Smile ID attribution or not on the Instructions screen
  * @param showInstructions Whether to deactivate capture screen's instructions for SmartSelfie.
- * @param partnerParams Custom values specific to partners
+ * @param extraPartnerParams Custom values specific to partners
  * @param colorScheme The color scheme to use for the UI. This is passed in so that we show a Smile
  * ID branded UI by default, but allow the user to override it if they want.
  * @param typography The typography to use for the UI. This is passed in so that we show a Smile ID
@@ -66,7 +66,7 @@ fun SmileID.SmartSelfieEnrollment(
     allowAgentMode: Boolean = false,
     showAttribution: Boolean = true,
     showInstructions: Boolean = true,
-    partnerParams: ImmutableMap<String, String> = persistentMapOf(),
+    extraPartnerParams: ImmutableMap<String, String> = persistentMapOf(),
     colorScheme: ColorScheme = SmileID.colorScheme,
     typography: Typography = SmileID.typography,
     onResult: SmileIDCallback<SmartSelfieResult> = {},
@@ -80,7 +80,7 @@ fun SmileID.SmartSelfieEnrollment(
             allowAgentMode = allowAgentMode,
             showAttribution = showAttribution,
             showInstructions = showInstructions,
-            partnerParams = partnerParams,
+            extraPartnerParams = extraPartnerParams,
             onResult = onResult,
         )
     }
@@ -102,7 +102,7 @@ fun SmileID.SmartSelfieEnrollment(
  * front camera will be used.
  * @param showAttribution Whether to show the Smile ID attribution or not on the Instructions screen
  * @param showInstructions Whether to deactivate capture screen's instructions for SmartSelfie.
- * @param partnerParams Custom values specific to partners
+ * @param extraPartnerParams Custom values specific to partners
  * @param colorScheme The color scheme to use for the UI. This is passed in so that we show a Smile
  * ID branded UI by default, but allow the user to override it if they want.
  * @param typography The typography to use for the UI. This is passed in so that we show a Smile ID
@@ -117,7 +117,7 @@ fun SmileID.SmartSelfieAuthentication(
     allowAgentMode: Boolean = false,
     showAttribution: Boolean = true,
     showInstructions: Boolean = true,
-    partnerParams: ImmutableMap<String, String> = persistentMapOf(),
+    extraPartnerParams: ImmutableMap<String, String> = persistentMapOf(),
     colorScheme: ColorScheme = SmileID.colorScheme,
     typography: Typography = SmileID.typography,
     onResult: SmileIDCallback<SmartSelfieResult> = {},
@@ -131,7 +131,7 @@ fun SmileID.SmartSelfieAuthentication(
             allowAgentMode = allowAgentMode,
             showAttribution = showAttribution,
             showInstructions = showInstructions,
-            partnerParams = partnerParams,
+            extraPartnerParams = extraPartnerParams,
             onResult = onResult,
         )
     }
@@ -164,7 +164,7 @@ fun SmileID.SmartSelfieAuthentication(
  * @param allowGalleryUpload Whether to allow the user to upload images from their gallery or not
  * @param showInstructions Whether to deactivate capture screen's instructions for Document
  * Verification (NB! If instructions are disabled, gallery upload won't be possible)
- * @param partnerParams Custom values specific to partners
+ * @param extraPartnerParams Custom values specific to partners
  * @param colorScheme The color scheme to use for the UI. This is passed in so that we show a Smile
  * ID branded UI by default, but allow the user to override it if they want.
  * @param typography The typography to use for the UI. This is passed in so that we show a Smile ID
@@ -185,7 +185,7 @@ fun SmileID.DocumentVerification(
     allowAgentMode: Boolean = false,
     allowGalleryUpload: Boolean = false,
     showInstructions: Boolean = true,
-    partnerParams: ImmutableMap<String, String> = persistentMapOf(),
+    extraPartnerParams: ImmutableMap<String, String> = persistentMapOf(),
     colorScheme: ColorScheme = SmileID.colorScheme,
     typography: Typography = SmileID.typography,
     onResult: SmileIDCallback<DocumentVerificationResult> = {},
@@ -212,7 +212,7 @@ fun SmileID.DocumentVerification(
                         documentType = documentType,
                         captureBothSides = captureBothSides,
                         selfieFile = bypassSelfieCaptureWithFile,
-                        partnerParams = partnerParams,
+                        extraPartnerParams = extraPartnerParams,
                     )
                 },
             ),
@@ -246,7 +246,7 @@ fun SmileID.DocumentVerification(
  * @param allowGalleryUpload Whether to allow the user to upload images from their gallery or not
  * @param showInstructions Whether to deactivate capture screen's instructions for Document
  * Verification (NB! If instructions are disabled, gallery upload won't be possible)
- * @param partnerParams Custom values specific to partners
+ * @param extraPartnerParams Custom values specific to partners
  * @param colorScheme The color scheme to use for the UI. This is passed in so that we show a Smile
  * ID branded UI by default, but allow the user to override it if they want.
  * @param typography The typography to use for the UI. This is passed in so that we show a Smile ID
@@ -266,7 +266,7 @@ fun SmileID.EnhancedDocumentVerificationScreen(
     allowAgentMode: Boolean = false,
     allowGalleryUpload: Boolean = false,
     showInstructions: Boolean = true,
-    partnerParams: ImmutableMap<String, String> = persistentMapOf(),
+    extraPartnerParams: ImmutableMap<String, String> = persistentMapOf(),
     colorScheme: ColorScheme = SmileID.colorScheme,
     typography: Typography = SmileID.typography,
     onResult: SmileIDCallback<EnhancedDocumentVerificationResult> = {},
@@ -292,7 +292,7 @@ fun SmileID.EnhancedDocumentVerificationScreen(
                         countryCode = countryCode,
                         documentType = documentType,
                         captureBothSides = captureBothSides,
-                        partnerParams = partnerParams,
+                        extraPartnerParams = extraPartnerParams,
                     )
                 },
             ),
@@ -321,7 +321,7 @@ fun SmileID.EnhancedDocumentVerificationScreen(
  * camera will be used.
  * @param showAttribution Whether to show the Smile ID attribution or not on the Instructions screen
  * @param showInstructions Whether to deactivate capture screen's instructions for SmartSelfie.
- * @param partnerParams Custom values specific to partners
+ * @param extraPartnerParams Custom values specific to partners
  * @param colorScheme The color scheme to use for the UI. This is passed in so that we show a Smile
  * ID branded UI by default, but allow the user to override it if they want.
  * @param typography The typography to use for the UI. This is passed in so that we show a Smile ID
@@ -337,7 +337,7 @@ fun SmileID.BiometricKYC(
     allowAgentMode: Boolean = false,
     showAttribution: Boolean = true,
     showInstructions: Boolean = true,
-    partnerParams: ImmutableMap<String, String> = persistentMapOf(),
+    extraPartnerParams: ImmutableMap<String, String> = persistentMapOf(),
     colorScheme: ColorScheme = SmileID.colorScheme,
     typography: Typography = SmileID.typography,
     onResult: SmileIDCallback<BiometricKycResult> = {},
@@ -351,7 +351,7 @@ fun SmileID.BiometricKYC(
             allowAgentMode = allowAgentMode,
             showAttribution = showAttribution,
             showInstructions = showInstructions,
-            partnerParams = partnerParams,
+            extraPartnerParams = extraPartnerParams,
             onResult = onResult,
         )
     }