+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Bump Sentry to 6.33.1
 - Bump Coil to 2.5.0
 - Changed the OKHTTP call timeout to 60 seconds 
+- Rename `partnerParams` to `extraPartnerParams`
 
 ### Removed